add with-span option :post-directly-p
There are situations where a span is posted to the queue, but then
somehow fails to make it to the HC servers. This needs more analysis.

As a workaround there's now the WITH-SPAN option :POST-DIRECTLY-P
which will synchronously post the span, meaning the WITH-SPAN form
will take a bit longer to complete.

Change-Id: Iff04357d211db46995d2aa9310a5f2288e6c138f
Reviewed-on: https://gerrit.franz.com:9080/c/cl-honeycomb/+/17203
Reviewed-by: Willem Broekema <metawilm@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Tested-by: John Foderaro <jkf@franz.com>